renamed home.js to index.js and made home page changes and mobile adjustments for form on calendar page
